v1.0.4  Replaced marked with commonmark.
Commonmark seemed a more up-to-date setup with a nice tree to go over the parsed noes. I hope it is a good idea. Anyway, replaced it and all the tests work now. The backslashes are a bit different. In particular backslashing punctuation will return that punctuation. So to backslash an underscore, you need two backslashes. Same with commas and colons. Hey, these were in my tests.

Commonmark does not fail when a link fails to conform. So the failure test is not relevant anymore Still thinking on how to alert about a malformed link syntax.