v1.1.0  Implemented better reporting
Reporting is much improved with human readable output. Hopefully sufficient. Folder.reporters along with appropriate emit messages allows for custom reporting of problems, mainly from commands or directives. Still a bit tricky.

Also fixed a problem with substitution in arguments; it was not appropriately checking for advancement to next argument.