start pipe directive
The pipe directive is an attempt to allow for a generalized mechanism of external communications that can easily take in something, run it through commands, and send out somewhere. It has a companion pipes also under development to allow it for directory-like objects as well.

This is almost at first draft stage.